cvs update
git-svn-id: svn+ssh://svn.berlios.de/svnroot/repos/ar7-firmware/qemu/trunk@363 11fd3d49-aaf6-0310-8052-ea9566248e5e

+++ hw/mips_r4k.c
@@ -15,7 +15,7 @@
 #define BIOS_FILENAME "mips_bios.bin"
 //#define BIOS_FILENAME "system.bin"
 #ifdef MIPS_HAS_MIPS64
-#define INITRD_LOAD_ADDR (int64_t)0x80800000
+#define INITRD_LOAD_ADDR (int64_t)(int32_t)0x80800000
 #else
 #define INITRD_LOAD_ADDR (int32_t)0x80800000
 #endif
